Move RequisitionRefusalException to its own file. Consolidate implementations of MeasurementResults
