+++ src/main/kotlin/org/wfanet/measurement/dataprovider/MeasurementResults.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 The Cross-Media Measurement Authors
+ * Copyright 2025 The Cross-Media Measurement Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,9 +14,12 @@
  * limitations under the License.
  */
 
-package org.wfanet.measurement.loadtest.dataprovider
+package org.wfanet.measurement.dataprovider
 
 import com.google.protobuf.TypeRegistry
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.collect
+import kotlinx.coroutines.flow.fold
 import org.projectnessie.cel.Program
 import org.wfanet.measurement.populationdataprovider.PopulationInfo
 import org.wfanet.measurement.populationdataprovider.PopulationRequisitionFulfiller
@@ -24,6 +27,36 @@ import org.wfanet.measurement.populationdataprovider.PopulationRequisitionFulfil
 /** Utilities for computing Measurement results. */
 object MeasurementResults {
   data class ReachAndFrequency(val reach: Int, val relativeFrequencyDistribution: Map<Int, Double>)
+
+  /**
+   * Computes reach and frequency using the "deterministic count distinct" methodology and the
+   * "deterministic distribution" methodology.
+   */
+  suspend fun computeReachAndFrequency(filteredVids: Flow<Long>, maxFrequency: Int): ReachAndFrequency {
+    // Count occurrences of each VID using fold operation on the flow
+    val eventsPerVid = filteredVids.fold(mutableMapOf<Long, Int>()) { acc, vid ->
+      acc[vid] = acc.getOrDefault(vid, 0) + 1
+      acc
+    }
+
+    val reach: Int = eventsPerVid.keys.size
+
+    // If the filtered VIDs is empty, set the distribution with all 0s up to maxFrequency.
+    if (reach == 0) {
+      return ReachAndFrequency(reach, (1..maxFrequency).associateWith { 0.0 })
+    }
+
+    // Build frequency histogram as a 0-based array.
+    val frequencyArray = IntArray(maxFrequency)
+    for (count in eventsPerVid.values) {
+      val bucket = count.coerceAtMost(maxFrequency)
+      frequencyArray[bucket - 1]++
+    }
+
+    val frequencyDistribution: Map<Int, Double> =
+      frequencyArray.withIndex().associateBy({ it.index + 1 }, { it.value.toDouble() / reach })
+    return ReachAndFrequency(reach, frequencyDistribution)
+  }
 
   /**
    * Computes reach and frequency using the "deterministic count distinct" methodology and the
@@ -51,8 +84,32 @@ object MeasurementResults {
   }
 
   /** Computes reach using the "deterministic count distinct" methodology. */
+  suspend fun computeReach(filteredVids: Flow<Long>): Int {
+    // Use a mutable set to track distinct VIDs as they flow through
+    val distinctVids = mutableSetOf<Long>()
+
+    filteredVids.collect { vid ->
+      distinctVids.add(vid)
+    }
+
+    return distinctVids.size
+  }
+
+  /** Computes reach using the "deterministic count distinct" methodology. */
   fun computeReach(filteredVids: Iterable<Long>): Int {
     return filteredVids.distinct().size
+  }
+
+  /** Computes impression using the "deterministic count" methodology. */
+  suspend fun computeImpression(filteredVids: Flow<Long>, maxFrequency: Int): Long {
+    // Count occurrences of each VID using fold operation on the flow
+    val eventsPerVid = filteredVids.fold(mutableMapOf<Long, Int>()) { acc, vid ->
+      acc[vid] = acc.getOrDefault(vid, 0) + 1
+      acc
+    }
+    
+    // Cap each count at `maxFrequency`.
+    return eventsPerVid.values.sumOf { count -> count.coerceAtMost(maxFrequency).toLong() }
   }
 
   /** Computes impression using the "deterministic count" methodology. */
@@ -70,4 +127,4 @@ object MeasurementResults {
   ): Long {
     return PopulationRequisitionFulfiller.computePopulation(populationInfo, program, typeRegistry)
   }
-}
+}
