+++ src/main/kotlin/org/wfanet/measurement/loadtest/dataprovider/BUILD.bazel
@@ -99,23 +99,13 @@ kt_jvm_library(
 )
 
 kt_jvm_library(
-    name = "measurement_results",
-    srcs = ["MeasurementResults.kt"],
-    deps = [
-        "//imports/java/org/projectnessie/cel",
-        "//src/main/kotlin/org/wfanet/measurement/populationdataprovider:population_requisition_fulfiller",
-        "@wfa_common_jvm//imports/java/com/google/protobuf",
-    ],
-)
-
-kt_jvm_library(
     name = "edp_simulator",
     srcs = ["EdpSimulator.kt"],
     deps = [
         ":event_query",
         ":frequency_vector_generator",
-        ":measurement_results",
         ":sketch_encrypter",
+        "//src/main/kotlin/org/wfanet/measurement/edpaggregator/resultsfulfiller:measurement_results",
         ":sketch_generator",
         ":vid_to_index_map_generator",
         "//src/main/kotlin/org/wfanet/measurement/api/v2alpha:packed_messages",
